Initial commit
This yaml builds the all four buckets, all four policies, all
four distributions, and all four Route53 records, and handles
switching between the primary site url subdomains and the
project-specific subdomains (assuming that the project will
start before it is named). The subdomains in it right now
are just for testing, they'll be changed to talusanalytics.com
once all of this works.